druuid v0.1.2 Druuid

Druuid is used to generate datetime relative uuids that fit into a postgres bigint. It does this by reserving the higher order bits for the datetime and the lower order bits for some entropy.

Getting Started

The primary function is Druuid.gen/1. This will return a new uuid based on the 1970 UNIX epoch:

Druuid.gen()
#=> 12283551831556210035

You may wish to choose an epoch offset to make this number smaller if you don’t need to worry about dates before a certain point. Druuid.epoch_offset/1 allows you to pass in an erlang datetime tuple as the offset point and Druuid.gen/1 can take this as an optional argument. For example, consider using Jan 1st, 2016 at midnight as an anchor point:

{{2016,1,1},{0,0,0}}
|> Druid.epoch_offset
|> Druuid.gen
#=> 106596357117955988

Functions

epoch_offset(offset_datetime)

Specs

epoch_offset(Tuple) :: integer

Calculates an epoch offset from the given datetime.

Parameters

offset_datetime erlang datetime tuple from which to calculate the offset

Examples

iex> Druuid.epoch_offset({{2016, 1, 1}, {0, 0, 0}})
1451606400

gen(epoch_offset \\ 0)

Specs

gen(integer) :: integer

Generates a druuid id given an epoch_offset (optional).

Parameters

epoch_offset (optional) integer value that offsets the 1970 epoch. Defaults to 0.

gen_from_values(epoch_offset, rand, ts)

Specs

gen_from_values(integer, float, integer) :: integer

This function determinstically generates the druuid id from the variables given. Unless you have a specific reason to override one of these, you probably want to use the Druuid.gen/1 function.

Parameters

epoch_offset integer value that offsets the 1970 UNIX epoch. Defaults to 0.

rand float value representing a random sample b/w 0 and 1 from a uniform distribution.

ts integer timestamp representing the current time in seconds from the epoch.

Examples

iex> Druuid.gen_from_values(0, 0.0, 1)
8388608000